Shuffle symbols from their current homes to better ones.
