+++ src/music.h
@@ -11,6 +11,10 @@
 #include <lua.h>
 
 
+#define MUSIC_FADEOUT_DELAY   1000 /**< Time it takes to fade out. */
+#define MUSIC_FADEIN_DELAY    2000 /**< Time it takes to fade in. */
+
+
 extern int music_disabled;
