+++ src/nopenal.h
@@ -6,6 +6,7 @@
 #ifndef NOPENAL_H
 #  define NOPENAL_H
 
+#include "log.h"
 #include "ncompat.h"
 
 #include "AL/alc.h"
@@ -104,5 +105,13 @@ extern ALvoid (AL_APIENTRY *nalEffecti)(ALuint,ALenum,ALint);
 extern ALvoid (AL_APIENTRY *nalEffectiv)(ALuint,ALenum,ALint*);
 extern ALvoid (AL_APIENTRY *nalEffectf)(ALuint,ALenum,ALfloat);
 extern ALvoid (AL_APIENTRY *nalEffectfv)(ALuint,ALenum,ALfloat*);
+
+/* Debugging.  */
+#ifdef DEBUGGING
+#define al_checkErr()      al_checkHandleError( __func__ )
+void al_checkHandleError( const char *func );
+#else /* DEBUG */
+#define al_checkErr() /**< Hack to ignore errors when debugging. */
+#endif /* DEBUG */
 
 #endif /* NOPENAL_H */
