+++ src/sound_openal.c
@@ -56,10 +56,6 @@
 #define SOUND_FADEOUT         100
 
 
-#define soundLock()     SDL_mutexP(sound_lock)
-#define soundUnlock()   SDL_mutexV(sound_lock)
-
-
 /*
  * Global sound lock.
  */
@@ -1468,49 +1464,3 @@ void sound_al_update (void)
       }
    }
 }
-
-
-#ifdef DEBUGGING
-/**
- * @brief Converts an OpenAL error to a string.
- *
- *    @param err Error to convert to string.
- *    @return String corresponding to the error.
- */
-void al_checkHandleError( const char *func )
-{
-   ALenum err;
-   const char *errstr;
-
-   /* Get the possible error. */
-   err = alGetError();
-
-   /* No error. */
-   if (err == AL_NO_ERROR)
-      return;
-
-   /* Get the message. */
-   switch (err) {
-      case AL_INVALID_NAME:
-         errstr = _("a bad name (ID) was passed to an OpenAL function");
-         break;
-      case AL_INVALID_ENUM:
-         errstr = _("an invalid enum value was passed to an OpenAL function");
-         break;
-      case AL_INVALID_VALUE:
-         errstr = _("an invalid value was passed to an OpenAL function");
-         break;
-      case AL_INVALID_OPERATION:
-         errstr = _("the requested operation is not valid");
-         break;
-      case AL_OUT_OF_MEMORY:
-         errstr = _("the requested operation resulted in OpenAL running out of memory");
-         break;
-
-      default:
-         errstr = _("unknown error");
-         break;
-   }
-   WARN(_("OpenAL error [%s]: %s"), func, errstr);
-}
-#endif /* DEBUGGING */
